Allowing coexistence of the operation and reference EOIAM IdP
