+++ vires_oauth/vires_oauth/providers/eoiam_ref/provider.py
@@ -1,10 +1,10 @@
 #-------------------------------------------------------------------------------
 #
-#  EOIAM provider - views
+#  Reference EOIAM provider - "social account" provider class
 #
 # Authors: Martin Paces <martin.paces@eox.at>
 #-------------------------------------------------------------------------------
-# Copyright (C) 2021 EOX IT Services GmbH
+# Copyright (C) 2023 EOX IT Services GmbH
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -26,26 +26,17 @@
 #-------------------------------------------------------------------------------
 # pylint: disable=missing-docstring
 
-#
-# Required settings:
-#
-# SOCIALACCOUNT_PROVIDERS = {
-#     'eoiam': {
-#         'SERVER_URL': <EOIAM server URL>,
-#     },
-# }
-
-from allauth.socialaccount import app_settings
-from allauth.socialaccount.providers.oauth2.views import (
-    OAuth2CallbackView, OAuth2LoginView,
-)
-from .provider import EoiamProvider
-from .views_base import EoiamOAuth2AdapterBase
+from ..eoiam.provider_base import EoiamProviderBase, extract_eoiam
+from ...forms import SignupForm
 
 
-class EoiamOAuth2Adapter(EoiamOAuth2AdapterBase):
-    provider_id = EoiamProvider.id
-    settings = app_settings.PROVIDERS.get(provider_id, {})
+class EoiamRefProvider(EoiamProviderBase):
+    id = "eoiam_ref"
+    logger_name = __name__
 
-oauth2_login = OAuth2LoginView.adapter_view(EoiamOAuth2Adapter)
-oauth2_callback = OAuth2CallbackView.adapter_view(EoiamOAuth2Adapter)
+provider_classes = [EoiamRefProvider]
+
+
+SignupForm.extractor(EoiamRefProvider.id)(extract_eoiam)
+
+
