+++ vires_oauth/vires_oauth/providers/eoiam_ref/views.py
@@ -1,10 +1,10 @@
 #-------------------------------------------------------------------------------
 #
-#  EOIAM provider - views
+#  Reference EOIAM provider - views
 #
 # Authors: Martin Paces <martin.paces@eox.at>
 #-------------------------------------------------------------------------------
-# Copyright (C) 2021 EOX IT Services GmbH
+# Copyright (C) 2023 EOX IT Services GmbH
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -40,12 +40,12 @@ from allauth.socialaccount.providers.oauth2.views import (
     OAuth2CallbackView, OAuth2LoginView,
 )
 from .provider import EoiamProvider
-from .views_base import EoiamOAuth2AdapterBase
+from ..eoiam.views_base import EoiamOAuth2AdapterBase
 
 
-class EoiamOAuth2Adapter(EoiamOAuth2AdapterBase):
+class EoiamRefOAuth2Adapter(EoiamOAuth2AdapterBase):
     provider_id = EoiamProvider.id
     settings = app_settings.PROVIDERS.get(provider_id, {})
 
-oauth2_login = OAuth2LoginView.adapter_view(EoiamOAuth2Adapter)
-oauth2_callback = OAuth2CallbackView.adapter_view(EoiamOAuth2Adapter)
+oauth2_login = OAuth2LoginView.adapter_view(EoiamRefOAuth2Adapter)
+oauth2_callback = OAuth2CallbackView.adapter_view(EoiamRefOAuth2Adapter)
